Extract ui module
