+++ acme_bot/music/__init__.py
@@ -22,14 +22,15 @@ from itertools import chain
 from random import choices
 from shutil import which
 
-from discord import Embed, ButtonStyle, ui
+from discord import Embed
 from discord.ext import commands
 from yt_dlp import YoutubeDL
 
 from acme_bot.autoloader import CogFactory, autoloaded
 from acme_bot.config.properties import MUSIC_EXTRACTOR_MAX_WORKERS
-from acme_bot.music.extractor import MusicExtractor, add_expire_time
-from acme_bot.music.player import MusicPlayer, PlayerState
+from acme_bot.music.extractor import MusicExtractor
+from acme_bot.music.player import MusicPlayer
+from acme_bot.music.ui import ConfirmAddTracks, SelectTrack
 
 log = logging.getLogger(__name__)
 
@@ -64,101 +65,6 @@ def export_entry_list(*iterables):
 def strip_urls(urls):
     """Strip entry strings from their title and duration, leaving the URL."""
     return (line.split()[0] for line in urls.split("\n") if line)
-
-
-class ConfirmAddTracks(ui.View):
-    """Confirm/cancel menu view for the play-url command."""
-
-    ACTION_TIMEOUT = 60.0
-
-    def __init__(self, user, player, results):
-        super().__init__(timeout=self.ACTION_TIMEOUT)
-        self.__user = user
-        self.__player = player
-        self.__results = results
-
-    async def interaction_check(self, interaction):
-        return interaction.user == self.__user
-
-    @ui.button(label="Add to queue", emoji="\u2795", style=ButtonStyle.primary)
-    async def add_to_queue(self, interaction, _):
-        """Confirm adding tracks to the player."""
-        await interaction.message.edit(
-            content=f"\u2795 {len(self.__results)} tracks added to the queue.",
-            delete_after=None,
-            view=None,
-        )
-
-        async with self.__player as player:
-            for track in self.__results:
-                add_expire_time(track)
-            player.extend(self.__results)
-
-            if player.state == PlayerState.IDLE:
-                await player.start_player(self.__results[0])
-
-    @ui.button(label="Cancel", emoji="\U0001F6AB", style=ButtonStyle.secondary)
-    async def cancel(self, interaction, _):
-        """Cancel adding tracks to the player."""
-        await interaction.message.edit(
-            content="\U0001F6AB *Action cancelled.*", view=None
-        )
-
-
-class SelectTrack(ui.View):
-    """Select menu view for the play/play-snd command."""
-
-    ACTION_TIMEOUT = 60.0
-
-    def __init__(self, user, player, return_queue, results):
-        super().__init__(timeout=self.ACTION_TIMEOUT)
-
-        self.__user = user
-        self.__player = player
-        self.__return_queue = return_queue
-
-        for index, new in enumerate(results, start=1):
-            self.add_select_button(index, new)
-        self.add_cancel_button()
-
-    async def interaction_check(self, interaction):
-        return interaction.user == self.__user
-
-    def add_select_button(self, index, new):
-        """Create a button that adds the given track to the player."""
-
-        async def button_pressed(interaction):
-            add_expire_time(new)
-            async with self.__player as player:
-                player.append(new)
-
-                if player.state == PlayerState.IDLE:
-                    msg = "\u2795 **{title}** by {uploader} added to the queue.".format(
-                        **new
-                    )
-                    await player.start_player(new)
-                    await interaction.message.edit(
-                        content=msg, delete_after=None, view=None
-                    )
-            await self.__return_queue.put(new)
-
-        button = ui.Button(label=str(index), style=ButtonStyle.secondary)
-        button.callback = button_pressed
-        self.add_item(button)
-
-    def add_cancel_button(self):
-        """Cancel adding tracks to the player."""
-
-        async def button_pressed(interaction):
-            await interaction.message.edit(
-                content="\U0001F6AB *Action cancelled.*", view=None
-            )
-
-        button = ui.Button(
-            label="Cancel", emoji="\U0001F6AB", style=ButtonStyle.secondary
-        )
-        button.callback = button_pressed
-        self.add_item(button)
 
 
 @autoloaded
